Mip Model

FlowGenerator.MipModel.CommodityFlowComponent — Type

struct CommodityFlowComponent

A struct for representing the flow component of a specific commodity within a network flow model.

Contains a JuMP implementation of:

the arc and path flow variables and flow conservation constraints associated with the commodity;

the flow demand and capacity constraints of the commodity.

FlowGenerator.MipModel.NetworkFlowMipModel — Type

struct NetworkFlowMipModel

A struct for representing a mixed-integer programming (MIP) model for network flow problems.

Fields

mip_model::JuMP.Model: The JuMP model object for the MIP.

problem::NetworkFlowModel.Problem: The network flow problem instance.

commodity_to_flow_data::Dict{Commodity,CommodityFlowComponent}: A dictionary mapping commodities to their respective flow components.

arc_to_capacity_constraint::Dict{NetworkFlowModel.Arc,JuMP.ConstraintRef}: A dictionary mapping arcs to their capacity constraints in the MIP model.

side_constrs::Vector{JuMP.ConstraintRef}: A vector of additional side constraints added to the MIP model.

dual_decimal_precision::Int: The precision used for the dual values retrieved from the JuMP model.

This struct is used to encapsulate all the components necessary to define and solve a MIP for a network flow problem, including the model itself, the problem data, constraints, and settings for numerical precision.

FlowGenerator.MipModel.NetworkFlowMipModel — Method

NetworkFlowMipModel(
    problem::NetworkFlowModel.Problem, 
    params::MipSolverParams
) -> NetworkFlowMipModel

Create a new mixed-integer programming (MIP) model for a network flow problem.

Returns

NetworkFlowMipModel: A NetworkFlowMipModel instance with initialized MIP model, constraints, and variables.

FlowGenerator.AbstractSolver.solve — Method

AbstractSolver.solve(
    problem::NetworkFlowModel.Problem, 
    params::MipSolverParams;
    time_limit::Float64 = Inf
) -> PrimalSolution

Solve a network flow problem directly as a mixed-integer programming (MIP) model. Keyword parameter time_limit is given in seconds.

Returns

A PrimalSolution object representing the optimal solution.

FlowGenerator.MipModel.add_arc_var! — Method

add_arc_var!(
    flow_component::CommodityFlowComponent, arc::Arc, cost::Float64; var_type::VarType
)

Add a flow variable for a specified arc to the CommodityFlowComponent if it does not already exist.

Returns

var: The JuMP variable reference for the newly added arc variable. Returns nothing if the variable already exists.

This function creates a variable associated with a specific arc in the network, assigns it a cost coefficient in the objective function, and updates flow conservation and commodity flow constraints to include this variable.

FlowGenerator.MipModel.add_arc_var! — Method

add_arc_var!(
    model::NetworkFlowMipModel,
    commodity::Commodity,
    arc::Arc;
    cost = get_cost(model.problem, arc),
    var_type = CONTINUOUS,
) -> Bool

Add a variable to the MIP model representing the flow of a commodity through a specific arc. The variable is added only if the arc is in the network of the problem and an equivalent variable does not already exist.

Returns

true if the variable was successfully added, otherwise false.

FlowGenerator.MipModel.add_path_var! — Method

add_path_var!(flow_component::CommodityFlowComponent, path::Path, cost::Float64)

Add a flow variable for a specified path to the CommodityFlowComponent if it does not already exist.

Returns

var: The JuMP variable reference for the newly added path variable. Returns nothing if the variable already exists.

This function creates a variable associated with a specific path in the network, assigns it a cost coefficient in the objective function, and updates the demand and capacity constraints to include this variable.

FlowGenerator.MipModel.add_path_var! — Method

add_path_var!(
    model::NetworkFlowMipModel,
    commodity::Commodity,
    path::Path;
    cost = get_cost(model.problem, path),
) -> Bool

Add a variable to the MIP model representing the flow of a commodity along a specific path. The variable is added only if the path is in the network of the problem and an equivalent variable does not already exist.

Returns

true if the variable was successfully added, otherwise false.

FlowGenerator.MipModel.delete_arc_var! — Method

delete_arc_var!(flow_component::CommodityFlowComponent, arc::Arc)

Delete an arc variable from the CommodityFlowComponent by setting the upper bound to 0 and removing it from assocaited data structures.

Note: setting the variable upper bound to 0 is more efficient than effectively deleting the variable.

FlowGenerator.MipModel.delete_path_var! — Method

delete_path_var!(flow_component::CommodityFlowComponent, path::Path)

Delete a path variable from the CommodityFlowComponent by setting the upper bound to 0 and removing it from assocaited data structures.

Note: setting the variable upper bound to 0 is more efficient than deleting the variable.

FlowGenerator.MipModel.get_arc_flow_solution — Method

get_arc_flow_solution(flow_component::CommodityFlowComponent) -> ArcFlowSolution

Retrieve the flow solution for all arcs within the CommodityFlowComponent.

This function calculates the flow solution for each arc by rounding the value of the associated variable and aggregating the flow values from the path variables as well.

FlowGenerator.MipModel.get_dual_solution — Method

get_dual_solution(model::NetworkFlowMipModel) -> DualSolution

Retrieve the dual solution from a solved NetworkFlowMipModel. Only works if the model is an LP model (all variables are linear).

FlowGenerator.MipModel.get_flow_conservation_constraint! — Method

get_flow_conservation_constraint!(
    flow_component::CommodityFlowComponent, vertex::Vertex
)

Retrieve or create a flow conservation constraint for a specified vertex within the CommodityFlowComponent.

Returns

constraint::JuMP.ConstraintRef: The JuMP constraint reference for the flow conservation at the specified vertex.

This function ensures that there is a flow conservation constraint for each vertex. If the constraint does not exist, it is created with a default expression of 0.0 == 0.0.

FlowGenerator.MipModel.get_paths — Method

get_paths(flow_component::CommodityFlowComponent)

Return all paths for which variables have been created in the CommodityFlowComponent.

FlowGenerator.MipModel.get_paths — Method

get_paths(mip_model::NetworkFlowMipModel, commodity::Commodity) -> Vector{Path}

Get all paths associated with a variable for the given commodity.

FlowGenerator.MipModel.get_primal_solution — Method

get_primal_solution(model::NetworkFlowMipModel) -> PrimalSolution

Retrieve the primal solution from a solved MIP model

FlowGenerator.MipModel.optimize! — Method

optimize!(model::NetworkFlowMipModel; time_limit::Float64 = Inf)

Optimize the MIP model associated with a network flow problem by calling JuMP.optimize!(). time_limit is given in seconds.

Returns

nothing.

FlowGenerator.NetworkFlowModel.get_arcs — Method

get_arcs(flow_component::CommodityFlowComponent)

Return all arcs for which variables have been created in the CommodityFlowComponent.

FlowGenerator.NetworkFlowModel.get_arcs — Method

get_arcs(mip_model::NetworkFlowMipModel, commodity::Commodity) -> Vector{Arc}

Get all arcs associated with a variable for the given commodity.