Network Flow Solver

FlowGenerator.AbstractSolver.solve — Method

AbstractSolver.solve(
    problem::NetworkFlowModel.Problem,
    params::NetworkFlowSolverParams;
    initial_paths::Vector{Tuple{Commodity,Path}} = Tuple{Commodity,Path}[],
    initial_columns = MipModel.Column[
        MipModel.Column(problem, path, commodity) for (commodity, path) in initial_paths
    ],
    obj_cutoff::Number = params.obj_cutoff,
    max_num_branching_levels::Int,
) -> NetworkFlowModel.PrimalSolution

Solves a network flow model problem by column generation, branching, and reduced-cost variable-fixing. The function applies unbalanced branching to favor finding good quality primal solutions, up to a specified maximum number of branching levels.